Adicionadas abstrações dos componentes de UI

+++ impl/core/src/main/java/br/gov/frameworkdemoiselle/behave/runner/ui/Screen.java
@@ -5,7 +5,7 @@ import br.gov.frameworkdemoiselle.behave.runner.ui.base.BaseUI;
 public interface Screen extends BaseUI {
 
 	void waitText(String text);
-	
-	void waitText(String text, Integer timeout);	
-	
+
+	void waitText(String text, Long timeout);
+
 }
